+++ source/extensions/lightspeed.common/lightspeed/common/__init__.py
@@ -14,5 +14,3 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 """
-
-from .reference_edit import *  # noqa F401