+++ source/extensions/omni.flux.utils.dialog/omni/flux/utils/dialog/__init__.py
@@ -15,4 +15,8 @@
 * limitations under the License.
 """
 
-from .progress_popup import *  # noqa F401
+__all__ = ["ErrorPopup", "show_popup", "ProgressPopup"]
+
+from .error_popup import ErrorPopup
+from .progress_popup import ProgressPopup
+from .simple_popup import show_popup