+++ source/extensions/lightspeed.trex.asset_replacements.core.shared/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [3.1.2]
+### Changed
+- Updated dependency from lightspeed.error_popup.window to omni.flux.utils.dialog
+
 ## [3.1.1]
 ### Changed
 - Added ability for the `get_textures_from_material_path` function to take a shader prim as input directly