+++ source/extensions/lightspeed.trex.capture_tree.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.1]
+### Changed
+- Updated imports from lightspeed.error_popup.window to omni.flux.utils.dialog
+
 ## [1.0.0]
 ### Added
 - Created