+++ source/extensions/lightspeed.trex.logic.widget/docs/CHANGELOG.md
@@ -4,6 +4,10 @@ This document records all notable changes to the **lightspeed.trex.logic.widget*
 
 The format is based on [Keep a Changelog](https://keepachangelog.com). The project adheres to [Semantic Versioning](https://semver.org).
 
+## [1.0.1]
+### Changed
+- Updated dependency from lightspeed.error_popup.window to omni.flux.utils.dialog
+
 ## [1.0.0]
 ### Added
 - Created based on `omni.graph.window.generic-1.60.0` and customized for Remix Components