+++ source/extensions/lightspeed.trex.properties_pane.shared.mod_packaging.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.1]
+### Changed
+- Updated popup imports to use unified omni.flux.utils.dialog extension
+
 ## [1.3.0]
 ### Added
 - Added workspace integration for dockable Mod Packaging window