+++ source/extensions/omni.flux.bookmark_tree.model.usd/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.4.2]
+### Changed
+- Updated imports for renamed extension: omni.flux.popup.window to omni.flux.utils.dialog
+
 ## [1.4.1]
 ## Fixed
 - Fixed Test assets to large to work without LFS